Rename: move notice api in club.js to notice.js
